get method uses rest framework better
The get method was not using rest framework best practices, namely
mixins/view methods to list or retrieve objects. Thus, it was not
accounting for filters on the queryset (such as from search queries).

I updated it to take full advantage of these builtin methods using
the List and Retrieve mixins. Now queries work.