fix initial upload bug re file field in serializer
With the file field in the serializer, it seems that
the file was getting created on the inital chunk with
that chunk's content, then the chunk was appended to
the end of the new file, duplicating the data. Instead,
we do not want to append the chunk if it is a new upload.
We do need to make sure to set the offset field on the
upload instance so successive chunks are appended to correct
location in the file.