Pass through responses sent from on_completion()
The previous implementation discarded any DRF Responses returned from on_completion.
To return a DRF response, the developer would have to copy the entire _post function to 
override it, which is inefficient. To counter this, see if the on_completion function has a return
value and pass it on if it does.